moved JsonError to api package; introduced JsonErrors utility class
https://github.com/metasfresh/metasfresh/issues/5469
